+++ CHANGELOG.md
@@ -6,6 +6,13 @@
 
 - Sentry-CLI integration test action: support envelopes ([#58](https://github.com/getsentry/github-workflows/pull/58))
 
+### Dependencies
+
+- Bump `danger/danger-js` from v11.1.2 to v11.3.0 ([#59](https://github.com/getsentry/github-workflows/pull/59))
+- Bump CLI from v2.0.0 to v2.0.4 ([#32](https://github.com/getsentry/github-workflows/pull/32))
+  - [changelog](https://github.com/getsentry/sentry-cli/blob/master/CHANGELOG.md#204)
+  - [diff](https://github.com/getsentry/sentry-cli/compare/2.0.0...2.0.4)
+
 ## 2.7.0
 
 ### Features